refactor: port ServerAppTests to QtTests
